requirements.txt
pyserial
pynput
qrcode
pillow
cryptography
pyinstaller 